Fixed empty entry condition and line end

+++ Compare.py
@@ -14,13 +14,15 @@ for line in input1:
     for line2 in input2:
         line2 = line2.rstrip('\n')
         fields2 = line2.split(',')
-        if fields[(int(sys.argv[2]) - 1)] == fields2[(int(sys.argv[4]) - 1)]:
+        fields2 = filter(None, fields2)
+        if (fields[(int(sys.argv[2]) - 1)] == fields2[(int(sys.argv[4]) - 1)]) and (len(fields2) > 1):
             del(fields2[(int(sys.argv[4]) - 1)])
             fields3 = fields + fields2
+            print fields3
             for data in fields3:
                 output.write(data + ',')
             output.seek(-1,1)
             output.truncate()
-            output.write(os.linesep)
+            output.write("\n")
 output.close()
